requirements.txt
apache-airflow==2.7.1
boto3==1.28.46
psycopg2-binary==2.9.7
pandas==2.2.3
requests==2.31.0
python-dotenv==1.0.0